Refactored parameter names with _as typos
A lot of parameters on the methods in the mapdl_functions seem to have
the same error, which is an _ inserted before the characters 'as'. So,
'base' had become 'b_ase', or 'mass', 'm_ass'. They are not deliberate
because they are corect in the docstrings.

I've gone through and removed the rogue underscores wherever I've found them.