added typings to new methods
Added a chunk of element methods to the new class. Additionally, besides
the move, I've added typings to all of these methods. The custom types
are below:

MapdlInt - should be an integer: either as a string or an int: "1" or 1
MapdlFloat - should be a float: either as a string or a float: "1.0" or
1.0
Union[str, int] - means that the input can be a string or an int, but
the string might not be an int: e.g. "1" 1 or "ALL"
A more complex custom type might work better in this case.

Union[str, float] - Vice versa for floats