Added internal links to the en command
Added links from the en docstring to the emore method. Linking within the
same file is very easy and does not require complex paths, however
linking outside of it does. So, I've not done the link for shpp as its
location will probably change in the future and may produce a bug. This
would likely happen to a lot of things though! So maybe we should bite
the bullet and hope that pycharm's refactoring methods can cope with
changing this as well.