Issue/444 mapdl element commands (#446)
* Moved element methods to new class + new type

Moved element methods away from mapdl_functions to a new class
_MapdlElementCommands. Some methods may have been missed. Amended
`self.run` to remove possibility of returning int and factored it into
the element_commands methods. Created a dedicated mapdl_types file and
added the MapdlInt type, which can be str or int.

* changed multiple bare excepts to include Exception

* added typings to new methods

Added a chunk of element methods to the new class. Additionally, besides
the move, I've added typings to all of these methods. The custom types
are below:

MapdlInt - should be an integer: either as a string or an int: "1" or 1
MapdlFloat - should be a float: either as a string or a float: "1.0" or
1.0
Union[str, int] - means that the input can be a string or an int, but
the string might not be an int: e.g. "1" 1 or "ALL"
A more complex custom type might work better in this case.

Union[str, float] - Vice versa for floats

* clean up element_commands

* add missing docbuild req

* finished line length formatting + string literals

All line lengths are less than 72 characters (for docstrings) and 79
characters (for code) as per PEP8. Additionally converted all renaming
string formatting statements to f-literals.

* removed references to GUI

I've now removed as many references to the GUI as I could find,
including bits about the GUI, or thing = P, etc.

* add missing files

* Update ansys/mapdl/core/mapdl.py

Co-authored-by: natter1 <52835756+natter1@users.noreply.github.com>

* Added set of tests for parse_e+

Added tests for parse_e and parse_et that check the handling of the
input and that it behaves as would be expected.

* update req

* modified docstrings to remove sphinx warnings

* we need to index the rst file to do links!

* Added internal links to the en command

Added links from the en docstring to the emore method. Linking within the
same file is very easy and does not require complex paths, however
linking outside of it does. So, I've not done the link for shpp as its
location will probably change in the future and may produce a bug. This
would likely happen to a lot of things though! So maybe we should bite
the bullet and hope that pycharm's refactoring methods can cope with
changing this as well.

* refactor prep7 commands

* finalize doc API

* add solution docs

* remove references to _MapdlCommands

* fix codespell, subpackages

* add untracked

* escaped most, if not all, trailing underscores in docstrings

* add unsupported commands

* quitely => quietly

* add apdl section

* add apdl section in earnest

* ignore api directory

* add post1 to docs

* moved docs

* started refactor of commands

* misc fixes for spelling and testing

* move doc source to correct directory

* add missing commands

* added last of v182 commands

* finish merge

* remove extra files

* remove extra files

* prep for refactor

* move map and aux15

* fix tests

* move aux3

* move aux3

* move reduced

* move aux2, aux12, database

* add missing files

* moved apdl commands

* add missing import

* move post26

* move prep7

* fix setup find_packages

* move graphic, post1, and solution

* fix whitespace

* fix spelling

* multi-threaded docbuild

* remove autogen

* use single thread on doc build

* remove depricated commands

* fix display rst

* add EEXTRUDE

* move more commands out of mapdl_functions

* remove remainder of depricated commands

* final removal of mapdl_functions

Co-authored-by: Alex Kaszynski <akascap@gmail.com>
Co-authored-by: natter1 <52835756+natter1@users.noreply.github.com>