Refactored commands into multiple classes
These commits shouldn't change any of the experience of users when
coding (i.e. the API has not changed appreciably), unless people have
been using methods directly from the files, which is NOT recommended
anyway.

They are just a refactor to make the commands class methods so that
intellisense works now in PyCharm (which it does) and suggestions
no longer suggest arguments off-by-one. As explained in #506.

This commit will close #506.

all the tests pass when run under the new refactor.