improved get_details_of_els + added elnext + tests
Added ELNEXT and tests. Made get_details_of_elements work like new
get_details_of_nodes, but left off node_numbers for now because
onfortunately it turns out if the list of nodes is longer than a certain
amount it wraps around. Parsing this is not trivial so I'll leave as it
is for now and fix this method in the future.