Added vlnext + tests
Added vlnext as well as an example + tests. In the tests I have added
`cleared` to the query fixture and `mapdl` to make it behave acceptably.

I have made a special function for vlnext tests because there was no
obvious fixture to use and we didn't need another one just for this.