Feat/issue 504 next selected (#534)
* added ndnext method as well as made get_node_deets better

* added kpnext + tests

* fixed text error

* improved get_details_of_els + added elnext + tests

Added ELNEXT and tests. Made get_details_of_elements work like new
get_details_of_nodes, but left off node_numbers for now because
onfortunately it turns out if the list of nodes is longer than a certain
amount it wraps around. Parsing this is not trivial so I'll leave as it
is for now and fix this method in the future.

* added elnext example + fixes to other docstrings

* added lsnext + tests

* added ARNEXT + tests

* Added vlnext + tests

Added vlnext as well as an example + tests. In the tests I have added
`cleared` to the query fixture and `mapdl` to make it behave acceptably.

I have made a special function for vlnext tests because there was no
obvious fixture to use and we didn't need another one just for this.

* updated Query docstring to include *next commands

Co-authored-by: Alex Kaszynski <akascap@gmail.com>