Made integer bool and added return type
Not sure why the parameter "integer" was a string in the default but
treated as a bool in the code. So I've made it a bool.

Additionally it didn't make sense for it to have a default value because
there's an equal number of integers versus floats. It would be better to
have to specify which is being requested whenever the method is called.