Added support for CWD to contain spaces
the cwd command accepts directory paths with spaces when enclosed in a
pair of single quotes/apostrophes. However, it seems that it does NOT
support directory path's that otherwise contain single quotes. I have
added a note to this effect to the docstring.

Otherwise this commit enables the supplication of directory paths with
spaces in as well as two tests for this. The methods could (and perhaps
should) be extended to other path-related commands, but I'm not sure
which these are, and would welcome input from other people.

Otherwise, this commit should fix issue #641.