Use gRPC functionality for inline functions (#639)
* fix use gRPC functionality

* refactor component_queries

* Made integer bool and added return type

Not sure why the parameter "integer" was a string in the default but
treated as a bool in the code. So I've made it a bool.

Additionally it didn't make sense for it to have a default value because
there's an equal number of integers versus floats. It would be better to
have to specify which is being requested whenever the method is called.

* simplified param parsing and changed selection status docstring

* altered all component queries to use new changes

* renamed class to reflect new query-running responsibilities

* refactored everything now except for selection queries (inc docstrings

* refactored all remaining selection queries

* add edge case testing

* remove unused import

* cannot use mute on line_geometry fixture

Co-authored-by: jgd10 <jamesgordonderrick@gmail.com>