All tests pass, with minor corrections
All tests pass under the new pathlib.Path system, however, a couple did
need to be changed that no longer functioned as intended. One was
getting random letter combinations from the filepath and is now doing so
explicitly instead. The other was testing that the filename always
consisted of / exclusively. however, this is no longer necessary given
the use of pathlib.Path.