build: temporary avoid flit latest version
Temporary avoid using the latest version of flit due to issues when releasing to PyPI with https://github.com/pypa/gh-action-pypi-publish. The problem comes from the new metadata format used to release and the fact the version of `pypa/gh-action-pypi-publish` we are using is not compatible with is. However, using a newer version is a problem, see ansys/actions#687 for more information on the blocking changes.