Created self-contained apdl/pymapdl conversion ex
The examples conspicuously don't contain any examples demonstarting
conversion from APDL to PyMAPDL scripts. In this, I do just that.

This example does not download any files and uses a modified version of
vm45 that has been hardcoded in. Otherwise, the example does save this
file to a temp folder using tempfile, and then loads it again to do the
conversion.